requirements.txt
opencv-contrib-python==4.7.0.68
mediapipe==0.9.1.0
numpy==1.24.1
pyyaml==6.0